+++ mmdet3d/evaluation/functional/waymo_utils/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 
-from .prediction_kitti_to_waymo import KITTI2Waymo
+from .prediction_to_waymo import Prediction2Waymo
 
-__all__ = ['KITTI2Waymo']
+__all__ = ['Prediction2Waymo']
