+++ mmdet3d/visualization/local_visualizer.py
@@ -23,12 +23,12 @@ from mmdet.visualization import DetLocalVisualizer
 from .vis_utils import (proj_camera_bbox3d_to_img, proj_depth_bbox3d_to_img,
                         proj_lidar_bbox3d_to_img, to_depth_mode)
 
-# try:
-#     import open3d as o3d
-#     from open3d import geometry
-# except ImportError:
-#     raise ImportError(
-#         'Please run "pip install open3d" to install open3d first.')
+try:
+    import open3d as o3d
+    from open3d import geometry
+except ImportError:
+    raise ImportError(
+        'Please run "pip install open3d" to install open3d first.')
 
 
 @VISUALIZERS.register_module()
@@ -126,8 +126,7 @@ class Det3DLocalVisualizer(DetLocalVisualizer):
             del self.pcd
             del self.points_colors
 
-    def _initialize_o3d_vis(self,
-                            frame_cfg):  #-> o3d.visualization.Visualizer:
+    def _initialize_o3d_vis(self, frame_cfg) -> o3d.visualization.Visualizer:
         """Initialize open3d vis according to frame_cfg.
 
         Args:
